requirements.txt
python-telegram-bot[ext]>=20.0 # Use latest stable or adjust